refactored app level, in preparation for react-router prevented wiggle if user multi-taps 'section 1' early return
