+++ src/assets/styles/Animations.css
@@ -11,8 +11,8 @@
   opacity: 0;
   margin: 1em auto;
   transform: translateX(-100%);
-  transition: opacity 1s ease-in-out, background-color 1s ease-in-out,
-    transform 1s ease-in-out; /* Adjust the duration and easing function as needed */
+  transition: opacity 1s ease-in-out, background-color 0.5s ease-in-out,
+    transform 1s ease-in-out, box-shadow 0.5s ease-in-out; /* Adjust the duration and easing function as needed */
 }
 
 .scroll-to-top.animate {
